# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisJobCreate": {
        "additionalProperties": false,
        "properties": {
          "engine_mode": {
            "$ref": "#/components/schemas/EngineMode",
            "default": "hosted_baseline"
          }
        },
        "title": "AnalysisJobCreate",
        "type": "object"
      },
      "AnalysisJobCreated": {
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "replayed": {
            "title": "Replayed",
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisJobStatus"
          }
        },
        "required": [
          "job_id",
          "status",
          "replayed"
        ],
        "title": "AnalysisJobCreated",
        "type": "object"
      },
      "AnalysisJobStatus": {
        "enum": [
          "queued",
          "running",
          "completed",
          "failed"
        ],
        "title": "AnalysisJobStatus",
        "type": "string"
      },
      "AnalysisResult": {
        "additionalProperties": false,
        "properties": {
          "abstain_reasons": {
            "items": {
              "maxLength": 120,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 20,
            "title": "Abstain Reasons",
            "type": "array"
          },
          "barriers": {
            "items": {
              "$ref": "#/components/schemas/Barrier"
            },
            "maxItems": 30,
            "title": "Barriers",
            "type": "array"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "maxItems": 50,
            "title": "Evidence",
            "type": "array"
          },
          "missing_information": {
            "items": {
              "maxLength": 240,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 30,
            "title": "Missing Information",
            "type": "array"
          },
          "model_score": {
            "anyOf": [
              {
                "maximum": 1,
                "minimum": 0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Score"
          },
          "precursors": {
            "items": {
              "$ref": "#/components/schemas/Precursor"
            },
            "maxItems": 30,
            "title": "Precursors",
            "type": "array"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "references": {
            "items": {
              "$ref": "#/components/schemas/ReferenceCitation"
            },
            "maxItems": 30,
            "title": "References",
            "type": "array"
          },
          "relevant_rule_ids": {
            "items": {
              "$ref": "#/components/schemas/RuleId"
            },
            "title": "Relevant Rule Ids",
            "type": "array"
          },
          "review_status": {
            "enum": [
              "pending",
              "reviewed"
            ],
            "title": "Review Status",
            "type": "string"
          },
          "rule_scores": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "maximum": 1,
                      "minimum": 0,
                      "type": "number"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "propertyNames": {
                  "$ref": "#/components/schemas/RuleId"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Scores"
          },
          "rules_assessment": {
            "$ref": "#/components/schemas/RulesAssessment"
          },
          "schema_version": {
            "const": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "score_kind": {
            "$ref": "#/components/schemas/ScoreKind"
          },
          "sif_label": {
            "$ref": "#/components/schemas/SifLabel"
          },
          "summary": {
            "maxLength": 800,
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          },
          "unassessed_rule_ids": {
            "items": {
              "$ref": "#/components/schemas/RuleId"
            },
            "title": "Unassessed Rule Ids",
            "type": "array"
          }
        },
        "required": [
          "relevant_rule_ids",
          "rules_assessment",
          "unassessed_rule_ids",
          "schema_version",
          "sif_label",
          "abstain_reasons",
          "model_score",
          "score_kind",
          "rule_scores",
          "precursors",
          "barriers",
          "evidence",
          "missing_information",
          "summary",
          "references",
          "review_status",
          "provenance"
        ],
        "title": "AnalysisResult",
        "type": "object"
      },
      "AuxiliaryModel": {
        "additionalProperties": false,
        "properties": {
          "model_id": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Model Id",
            "type": "string"
          },
          "provider": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Provider",
            "type": "string"
          },
          "purpose": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Purpose",
            "type": "string"
          },
          "revision": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Revision",
            "type": "string"
          }
        },
        "required": [
          "purpose",
          "provider",
          "model_id",
          "revision"
        ],
        "title": "AuxiliaryModel",
        "type": "object"
      },
      "Barrier": {
        "additionalProperties": false,
        "properties": {
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Evidence Ids",
            "type": "array"
          },
          "state": {
            "$ref": "#/components/schemas/BarrierState"
          },
          "tag": {
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[a-z0-9_]+$",
            "title": "Tag",
            "type": "string"
          }
        },
        "required": [
          "tag",
          "state",
          "evidence_ids"
        ],
        "title": "Barrier",
        "type": "object"
      },
      "BarrierState": {
        "enum": [
          "failed",
          "missing",
          "present",
          "unknown"
        ],
        "title": "BarrierState",
        "type": "string"
      },
...
      "EngineMode": {
        "enum": [
          "hosted_baseline",
          "encoder",
          "qwen_adapter",
          "recorded_sample"
        ],
        "title": "EngineMode",
        "type": "string"
      },
      "Evidence": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "enum": [
              "narrative",
              "activity"
            ],
            "title": "Field",
            "type": "string"
          },
          "id": {
            "pattern": "^e[1-9][0-9]*$",
            "title": "Id",
            "type": "string"
          },
          "method": {
            "$ref": "#/components/schemas/EvidenceMethod"
          },
          "occurrence": {
            "default": 0,
            "minimum": 0,
            "title": "Occurrence",
            "type": "integer"
          },
          "quote": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Quote",
            "type": "string"
          }
        },
        "required": [
          "id",
          "field",
          "quote",
          "method"
        ],
        "title": "Evidence",
        "type": "object"
      },
      "EvidenceMethod": {
        "enum": [
          "llm_extracted",
          "rule_matched",
          "reviewer"
        ],
        "title": "EvidenceMethod",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "analysis_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Id"
          },
          "attempt_count": {
            "maximum": 2,
            "minimum": 0,
            "title": "Attempt Count",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "report_id": {
            "title": "Report Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisJobStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "report_id",
          "status",
          "attempt_count",
          "error_code",
          "analysis_id",
          "updated_at"
        ],
        "title": "JobStatusResponse",
        "type": "object"
      },
...
      "Precursor": {
        "additionalProperties": false,
        "properties": {
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Evidence Ids",
            "type": "array"
          },
          "tag": {
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[a-z0-9_]+$",
            "title": "Tag",
            "type": "string"
          }
        },
        "required": [
          "tag",
          "evidence_ids"
        ],
        "title": "Precursor",
        "type": "object"
      },
      "Provenance": {
        "additionalProperties": false,
        "properties": {
          "auxiliary_models": {
            "items": {
              "$ref": "#/components/schemas/AuxiliaryModel"
            },
            "maxItems": 5,
            "title": "Auxiliary Models",
            "type": "array"
          },
          "engine_mode": {
            "$ref": "#/components/schemas/EngineMode"
          },
          "inference_is_live": {
            "title": "Inference Is Live",
            "type": "boolean"
          },
          "label_guide_version": {
            "const": "sif-pilot-v1",
            "title": "Label Guide Version",
            "type": "string"
          },
          "model_id": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Model Id",
            "type": "string"
          },
          "model_revision": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Model Revision",
            "type": "string"
          },
          "prompt_version": {
            "const": "sif-analysis-v1",
            "title": "Prompt Version",
            "type": "string"
          }
        },
        "required": [
          "engine_mode",
          "model_id",
          "model_revision",
          "prompt_version",
          "label_guide_version",
          "auxiliary_models",
          "inference_is_live"
        ],
        "title": "Provenance",
        "type": "object"
      },
...
      "ReferenceCitation": {
        "additionalProperties": false,
        "properties": {
          "passage_id": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Passage Id",
            "type": "string"
          },
          "source_id": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Source Id",
            "type": "string"
          }
        },
        "required": [
          "source_id",
          "passage_id"
        ],
        "title": "ReferenceCitation",
        "type": "object"
      },
      "ReportCreate": {
        "additionalProperties": false,
        "properties": {
          "activity": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Activity"
          },
          "is_synthetic": {
            "default": false,
            "title": "Is Synthetic",
            "type": "boolean"
          },
          "narrative": {
            "maxLength": 8000,
            "minLength": 20,
            "title": "Narrative",
            "type": "string"
          },
          "report_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Report Date"
          },
          "site": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Site"
          }
        },
        "required": [
          "narrative"
        ],
        "title": "ReportCreate",
        "type": "object"
      },
      "ReportCreated": {
        "additionalProperties": false,
        "properties": {
          "replayed": {
            "title": "Replayed",
            "type": "boolean"
          },
          "report_id": {
            "title": "Report Id",
            "type": "string"
          },
          "version": {
            "exclusiveMinimum": 0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "report_id",
          "version",
          "replayed"
        ],
        "title": "ReportCreated",
        "type": "object"
      },
      "ReportDetail": {
        "additionalProperties": false,
        "properties": {
          "analyses": {
            "items": {
              "$ref": "#/components/schemas/StoredAnalysis"
            },
            "title": "Analyses",
            "type": "array"
          },
          "report": {
            "$ref": "#/components/schemas/ReportRecord"
          }
        },
        "required": [
          "report",
          "analyses"
        ],
        "title": "ReportDetail",
        "type": "object"
      },
      "ReportRecord": {
        "additionalProperties": false,
        "properties": {
          "activity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_synthetic": {
            "title": "Is Synthetic",
            "type": "boolean"
          },
          "narrative": {
            "title": "Narrative",
            "type": "string"
          },
          "report_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Date"
          },
          "report_group_id": {
            "title": "Report Group Id",
            "type": "string"
          },
          "site": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site"
          },
          "version": {
            "exclusiveMinimum": 0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "report_group_id",
          "version",
          "narrative",
          "activity",
          "site",
          "report_date",
          "is_synthetic",
          "created_at"
        ],
        "title": "ReportRecord",
        "type": "object"
      },
      "RuleId": {
        "enum": [
          "bypassing_safety_controls",
          "confined_space",
          "driving",
          "energy_isolation",
          "hot_work",
          "line_of_fire",
          "safe_mechanical_lifting",
          "work_authorisation",
          "working_at_height"
        ],
        "title": "RuleId",
        "type": "string"
      },
      "RulesAssessment": {
        "enum": [
          "complete",
          "partial",
          "unknown"
        ],
        "title": "RulesAssessment",
        "type": "string"
      },
      "ScoreKind": {
        "enum": [
          "none",
          "uncalibrated_model_score",
          "calibrated_score"
        ],
        "title": "ScoreKind",
        "type": "string"
      },
      "SifLabel": {
        "enum": [
          "yes",
          "no",
          "unknown"
        ],
        "title": "SifLabel",
        "type": "string"
      },
      "StoredAnalysis": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_ms": {
            "minimum": 0,
            "title": "Duration Ms",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/AnalysisResult"
          }
        },
        "required": [
          "id",
          "job_id",
          "result",
          "duration_ms",
          "created_at"
        ],
        "title": "StoredAnalysis",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Safeguard AI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Job"
      }
    },
    "/v1/reports": {
      "post": {
        "operationId": "create_report_v1_reports_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Report"
      }
    },
    "/v1/reports/{report_id}": {
      "get": {
        "operationId": "get_report_v1_reports__report_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Report"
      }
    },
    "/v1/reports/{report_id}/analyses": {
      "post": {
        "operationId": "create_analysis_job_v1_reports__report_id__analyses_post",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisJobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisJobCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Analysis Job"
      }
    }
  }
}